Move scrubbing code to separate file
